Tests/LibWeb/Text/input/HTML/shadow-realm-sync-evaluate.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const realm = new ShadowRealm();
        const result = realm.evaluate(`() => 1 + 1`)();
        println(`Shadow realm evaluation returned: ${result}`);
    });
</script>
