Tests/LibWeb/Text/expected/HTML/shadow-realm-sync-evaluate.txt

Shadow realm evaluation returned: 2
